+++ hostfact_collector.go
@@ -17,6 +17,11 @@ import (
 var (
 	hostsFactsKey          = "collectors/host-fact"
 	hostFactsCollectorLock = make(chan struct{}, 1)
+
+	hostFactScrapeDurationMetric = prometheus.NewGauge(prometheus.GaugeOpts{
+		Name: "foreman_exporter_host_facts_scrape_duration_seconds",
+		Help: "Duration of the last completed host facts collector scrape of foreman.",
+	})
 )
 
 type HostFactCollector struct {
@@ -124,6 +129,7 @@ func (c HostFactCollector) Collect(ch chan<- prometheus.Metric) {
 
 		// use a goroutine to get result in async mode
 		go func() {
+			start := time.Now()
 
 			if *collectorsLock {
 				// lock and return directly if another request is in progress
@@ -160,6 +166,9 @@ func (c HostFactCollector) Collect(ch chan<- prometheus.Metric) {
 					hostsData = append(hostsData, labels)
 				}
 
+				elapsed := time.Since(start)
+				hostFactScrapeDurationMetric.Set(elapsed.Seconds())
+
 				// Add to the cache
 				if c.RingConfig.enabled && c.CacheConfig.Enabled {
 					content, _ := json.Marshal(hostsData)
@@ -169,12 +178,12 @@ func (c HostFactCollector) Collect(ch chan<- prometheus.Metric) {
 					}
 					if hostsFactsError == nil {
 						// update the cache
-						c.Logger.Info(fmt.Sprintf("updating cache key '%s'", hostsFactsKey))
+						c.Logger.Info(fmt.Sprintf("updating cache key '%s'", hostsFactsKey), "duration", elapsed.String())
 						c.updateKV(content)
 					}
 				} else if c.CacheConfig.Enabled {
 					// update the local cache
-					c.Logger.Info(fmt.Sprintf("updating cache key '%s'", hostsFactsKey))
+					c.Logger.Info(fmt.Sprintf("updating cache key '%s'", hostsFactsKey), "duration", elapsed.String())
 					localCache.Set(hostsFactsKey, hostsData, c.CacheConfig.ExpiresTTL)
 				}
 			}